form-validation: add lib
add lib and first mockup to the project

form-data: add new form components

the customer and billing pages shouldn't have the proper form as we are
goint to separate them in 3 different parts, to mobile will use 2 of
them and the desktop version will only use one version with both forms
together

there're still some stuff to fix, but this commit add the first
separation

form-validation: remove formik

replacing with react-vanilla-form
