--- src/theme-pagarme/billing-page/properties.css
@@ -1,14 +1,0 @@
-@import "../colors.css";
-@import "../typography.css";
-@import "../spacing.css";
-@import "../media.css";
-@import "../margin.css";
-
-:root {
-  --page-margin-desktop: var(--spacing-medium) 0 var(--spacing-medium) var(--spacing-small);
-  --page-title-font-family: var(--primary-font-family);
-  --page-title-font-size: var(--font-size-small);
-  --page-title-color: var(--color-white);
-  --page-title-margin: 6px 0;
-  --page-title-padding-bottom: 24px;
-}
