+++ src/theme-pagarme/general-form/properties.css
